# tp redes giercowski gonzalez/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "API Videos",
        "description": "API para manejar videos y sus vistas"
        },
    "servers": [
    {
        "url": "http://localhost:1815"
    }
    ],
    "paths": {
        "/videos": {
            "get": {
                "summary": "Muestra los videos",
                "description": "Muestra todos los videos",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                    },
                                                "nombre": {
                                                    "type": "string"
                                                    },
                                                "duracion": {
                                                    "type": "integer"
                                                    },
                                                "miniatura": {
                                                    "type": "string"
                                                    },
                                                "usuario": {
                                                    "type": "string"
                                                }                                              
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "/videos/{id}/{titulo}/{duracion}/{miniatura}/{usuario}" : {
                "post": {
                    "summary": "Subir video",
                    "description": "Sube el video que reciba",
                    "parameters" : [{
                        "name" : "id",
                        "description": "Id de la vista",
                        "schema" : {
                            "type": "integer"
                        },
                        "in" : "path",
                        "required" : "true"
                    },{
                        "name" : "titulo",
                        "description": "titulo del video",
                        "schema" : {
                            "type": "string"
                        },
                        "in" : "path",
                        "required" : "true"
                    },{
                        "name" : "duracion",
                        "description": "Duración del video",
                        "schema" : {
                            "type": "integer"
                        },
                        "in" : "path",
                        "required" : "true"
                    },{
                        "name" : "miniatura",
                        "description": "miniatura del video, en base 64",
                        "schema" : {
                            "type": "string"
                        },
                        "in" : "path",
                        "required" : "true"
                    }, {
                        "name" : "usuario",
                        "description": "nombre de usuario de quien sube el video",
                        "schema" : {
                            "type": "string"
                        },
                        "in" : "path",
                        "required" : "true"
                    }],
                    "responses": {
                        "200": {
                            "description": "Operación exitosa",
                                "content": {
                                    "application/json": {
                                        "schema": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id vista": {
                                                        "type": "integer"
                                                        },
                                                    "id video": {
                                                        "type": "integer"
                                                        },                                  
                                                    "duracion": {
                                                        "type": "integer"
                                                        },
                                                    "ubicacion": {
                                                        "type": "string"
                                                        },
                                                    "fecha": {
                                                        "type": "string"
                                                    }                                              
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "put": {
                        "summary": "Modificar un video",
                        "description": "Sobreescribe los datos de un video con los enviados como parámetros",
                        "parameters" : [{
                            "name" : "id",
                            "description": "Id del video",
                            "schema" : {
                                "type": "integer"
                            },
                            "in" : "path",
                            "required" : "true"
                        },{
                            "name" : "titulo",
                            "description": "titulo del video",
                            "schema" : {
                                "type": "string"
                            },
                            "in" : "path",
                            "required" : "true"
                        },{
                            "name" : "duracion",
                            "description": "Duración del video",
                            "schema" : {
                                "type": "integer"
                            },
                            "in" : "path",
                            "required" : "true"
                        },{
                            "name" : "miniatura",
                            "description": "miniatura del video, en base 64",
                            "schema" : {
                                "type": "string"
                            },
                            "in" : "path",
                            "required" : "true"
                        }, {
                            "name" : "usuario",
                            "description": "nombre de usuario de quien sube el video",
                            "schema" : {
                                "type": "string"
                            },
                            "in" : "path",
                            "required" : "true"
                        }],
                        "responses": {
                            "200": {
                                "description": "Operación exitosa",
                                    "content": {
                                        "application/json": {
                                            "schema": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                            },
                                                        "nombre": {
                                                            "type": "string"
                                                            },
                                                        "duracion": {
                                                            "type": "integer"
                                                            },
                                                        "miniatura": {
                                                            "type": "string"
                                                            },
                                                        "usuario": {
                                                            "type": "string"
                                                        }                                              
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
            },
            "/videos/{id}" : {
            "delete": {
                "summary": "Elminar un video",
                "description": "Elimina el video cuyo id coincide con el recibido",
                "parameters" : [{
                    "name" : "id",
                    "description": "Id del video a borrar",
                    "schema" : {
                        "type" : "integer"
                    },
                    "in" : "path",
                    "required" : "true"
                }],
                "responses": {
                    "204": {
                        "description": "Operación exitosa",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                    },
                                                "nombre": {
                                                    "type": "string"
                                                    },
                                                "duracion": {
                                                    "type": "integer"
                                                    },
                                                "miniatura": {
                                                    "type": "string"
                                                    },
                                                "usuario": {
                                                    "type": "string"
                                                }                                              
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "/vistas/{id}/{idVideo}/{duracion}/{ubicacion}/{fecha}":{
                "post": {
                    "summary": "Realizar vista",
                    "description": "Agrega una vista al video indicado",
                    "parameters" : [{
                        "name" : "id",
                        "description": "Id de la vista",
                        "in" : "path",
                        "required" : "true",
                        "schema" : {
                            "type" : "integer"
                        }
                    },{
                        "name" : "idVideo",
                        "description": "Id del video",
                        "in" : "path",
                        "required" : "true",
                        "schema" : {
                            "type" : "string"
                        }
                    },{
                        "name" : "duracion",
                        "description": "Duración de la vista",
                        "in" : "path",
                        "required" : "true",
                        "schema" : {
                            "type" : "integer"
                        }
                    },{
                        "name" : "ubicacion",
                        "description": "Lugar desde donde se realizo la vista",
                        "in" : "path",
                        "required" : "true",
                        "schema" : {
                            "type" : "string"
                        }
                    }, {
                        "name" : "fecha",
                        "description": "Fecha en la que se realizo la vista",
                        "in" : "path",
                        "required" : "true",
                        "schema" : {
                            "type" : "string"
                        }
                    }],
                    "responses": {
                        "200": {
                            "description": "Operación exitosa",
                                "content": {
                                    "application/json": {
                                        "schema": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id vista": {
                                                        "type": "integer"
                                                        },
                                                    "id video": {
                                                        "type": "integer"
                                                        },                                  
                                                    "duracion": {
                                                        "type": "integer"
                                                        },
                                                    "ubicacion": {
                                                        "type": "string"
                                                        },
                                                    "fecha": {
                                                        "type": "string"
                                                    }                                              
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                        "put": {
                            "summary": "Modifica una vista",
                            "description": "Modifica un video con los datos obtenidos por parametros",
                            "parameters" : [{
                                "name" : "id",
                                "description": "Id de la vista",
                                "in" : "path",
                                "required" : "true",
                                "schema" : {
                                    "type" : "integer"
                                }
                            },{
                                "name" : "idVideo",
                                "description": "Id del video",
                                "in" : "path",
                                "required" : "true",
                                "schema" : {
                                    "type" : "string"
                                }
                            },{
                                "name" : "duracion",
                                "description": "Duración de la vista",
                                "in" : "path",
                                "required" : "true",
                                "schema" : {
                                    "type" : "integer"
                                }
                            },{
                                "name" : "ubicacion",
                                "description": "Lugar desde donde se realizo la vista",
                                "in" : "path",
                                "required" : "true",
                                "schema" : {
                                    "type" : "string"
                                }
                            }, {
                                "name" : "fecha",
                                "description": "Fecha en la que se realizo la vista",
                                "in" : "path",
                                "required" : "true",
                                "schema" : {
                                    "type" : "string"
                                }
                            }],
                            "responses": {
                                "200": {
                                    "description": "Operación exitosa",
                                        "content": {
                                            "application/json": {
                                                "schema": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id vista": {
                                                                "type": "integer"
                                                                },
                                                            "id video": {
                                                                "type": "integer"
                                                                },                                  
                                                            "duracion": {
                                                                "type": "integer"
                                                                },
                                                            "ubicacion": {
                                                                "type": "string"
                                                                },
                                                            "fecha": {
                                                                "type": "string"
                                                            }                                              
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
    
            },
            "/vistas/{id}" : {
                "delete": {
                    "summary": "Eliminar una vista",
                    "description": "Eliminar la vista con el id recibido",
                    "parameters" : [{
                        "name" : "id",
                        "description": "Id de la vista a borrar",
                        "schema" : {
                            "type" : "integer"
                        },
                        "required" : "true",
                        "in" : "path"
                    }],
                    "responses": {
                        "200": {
                            "description": "Operación exitosa",
                                "content": {
                                    "application/json": {
                                        "schema": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                        },
                                                    "nombre": {
                                                        "type": "string"
                                                        },
                                                    "duracion": {
                                                        "type": "integer"
                                                        },
                                                    "miniatura": {
                                                        "type": "string"
                                                        },
                                                    "usuario": {
                                                        "type": "string"
                                                    }                                              
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
            },
            "/verVideosXVistas/{cantMinima}":{
                "get":{
                    "summary": "Videos con cierta cantidad de visitas",
                    "description": "Muestra todos los videos con mas vistas que el numero recibido",
                    "parameters":[{
                        "name": "cantMinima",
                        "description": "cantidad minima de visitas para hacer la busqueda",
                        "schema" : {
                            "type" : "integer"
                        },
                        "required" : "true",
                        "in" : "path"
                    }],
                    "responses": {
                        "200": {
                            "description": "Operación exitosa",
                                "content": {
                                    "application/json": {
                                        "schema": {
                                            "type": "array",
                                            "items": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                    },
                                                "nombre": {
                                                    "type": "string"
                                                    },
                                                "duracion": {
                                                    "type": "integer"
                                                    },
                                                "miniatura": {
                                                    "type": "string"
                                                    },
                                                "usuario": {
                                                    "type": "string"
                                                }                                              
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "/verVistasVideos":{
                "get":{
                    "summary": "Mostrar videos",
                    "description": "Muestra todos los videos y su cantidad de visitas",
                    "responses": {
                        "200": {
                            "description": "Operación exitosa",
                                "content": {
                                    "application/json": {
                                        "schema": {
                                            "type": "array",
                                            "items": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                    },
                                                "nombre": {
                                                    "type": "string"
                                                    },
                                                "duracion": {
                                                    "type": "integer"
                                                    },
                                                "miniatura": {
                                                    "type": "string"
                                                    },
                                                "usuario": {
                                                    "type": "string"
                                                }                                              
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "/verVideoMasLargo":{
                "get":{
                    "summary": "Mostrar videos mas largos",
                    "description": "Muestra el o los videos con mayor duracion",
                    "responses": {
                        "200": {
                            "description": "Operación exitosa",
                                "content": {
                                    "application/json": {
                                        "schema": {
                                            "type": "array",
                                            "items": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                    },
                                                "nombre": {
                                                    "type": "string"
                                                    },
                                                "duracion": {
                                                    "type": "integer"
                                                    },
                                                "miniatura": {
                                                    "type": "string"
                                                    },
                                                "usuario": {
                                                    "type": "string"
                                                }                                              
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
    